Move some common logic for excellibs benchmarks to the SheetOperationsBenchmarkBase class.
